pyproject.toml
[project]
name = "orkflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-json-widget>=2.1.1",
    "djangorestframework>=3.16.1",
    "ipython>=9.8.0",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.8",
    "djangorestframework-stubs>=3.16.6",
    "ipython>=9.8.0",
]
